Add files via upload
# Predicting Customer Churn
Objective:
To train the model using existing datasets and predicting the likelihood that a customer will churn.

Customer churn prediction refers to detecting which customers are likely to leave a service or terminate their subscription to a service. It is very critical for businesses because gaining new clients often costs more than retaining the existing ones.

Customer churn is a common problem that exists across businesses in many industries. Investing in acquiring new clients is a non-negotiable investment for any company. When a client leaves, it represents a significant loss for the business. A lot of resources needs to be channeled into replacing them.

Some of the reasons why customers are lost include:

Incorrect pricing
Unimpressive renewal offers.
Lack of market understanding.
poor customer service
Overcommunication and spamming.
Lack of brand loyalty.
The ability to predict when a client is likely to leave and offering them enticing incentives that will make them stay, can offer huge savings to a business. Business managers have to understand the factors that keep customers engaged and this is a strategy that requires constant development. It is important to note that finding patterns using Exploratory Data Analysis (EDA) is as important as the final prediction itself. A Churn prediction task remains unfinished if the data patterns are not found in EDA.

This churn prediction modelling technique attempts to explore some customer behaviours and attributes that relates to the reason and time of customers leaving.